rescale cleanup
