+++ src/circuit/ops/lookup.rs
@@ -175,7 +175,9 @@ impl<F: PrimeField + TensorType + PartialOrd> Op<F> for LookupOp {
                     scale_to_multiplier(global_scale) as usize,
                 ),
             }),
-            LookupOp::GreaterThan { .. } => Box::new(self.clone()),
+            LookupOp::GreaterThan { a } => Box::new(LookupOp::GreaterThan {
+                a: utils::F32(a.0 * scale_to_multiplier(inputs_scale[0])),
+            }),
         }
     }
